+++ business/src/main/java/com/kevinguanchedarias/owgejava/repository/RuleRepository.java
@@ -12,6 +12,10 @@ public interface RuleRepository extends JpaRepository<Rule, Integer> {
 
     List<Rule> findByOriginTypeAndOriginId(String sourceType, long id);
 
+    boolean existsByOriginTypeAndOriginIdAndDestinationTypeIn(
+            String originType, Long originId, List<String> destinationType
+    );
+
     Optional<Rule> findOneByTypeAndOriginTypeAndOriginIdAndDestinationTypeAndDestinationId(
             String type, String originType, Long originId, String destinationType, Long destinationId
     );