+++ game-rest/src/main/java/com/kevinguanchedarias/owgejava/rest/game/UnitRestService.java
@@ -6,17 +6,14 @@ import com.kevinguanchedarias.owgejava.business.FactionBo;
 import com.kevinguanchedarias.owgejava.business.MissionBo;
 import com.kevinguanchedarias.owgejava.business.ObtainedUnitBo;
 import com.kevinguanchedarias.owgejava.business.RequirementBo;
-import com.kevinguanchedarias.owgejava.business.SpeedImpactGroupBo;
 import com.kevinguanchedarias.owgejava.business.UnitBo;
 import com.kevinguanchedarias.owgejava.business.UnlockedRelationBo;
 import com.kevinguanchedarias.owgejava.business.UserStorageBo;
-import com.kevinguanchedarias.owgejava.business.unit.util.UnitTypeInheritanceFinderService;
 import com.kevinguanchedarias.owgejava.dto.ObtainedUnitDto;
 import com.kevinguanchedarias.owgejava.dto.RunningUnitBuildDto;
 import com.kevinguanchedarias.owgejava.dto.UnitDto;
 import com.kevinguanchedarias.owgejava.entity.ObtainedUnit;
 import com.kevinguanchedarias.owgejava.entity.Unit;
-import com.kevinguanchedarias.owgejava.entity.UnitType;
 import com.kevinguanchedarias.owgejava.entity.UserStorage;
 import com.kevinguanchedarias.owgejava.enumerations.ObjectEnum;
 import com.kevinguanchedarias.owgejava.interfaces.SyncSource;
@@ -24,7 +21,6 @@ import com.kevinguanchedarias.owgejava.pojo.DeprecationRestResponse;
 import com.kevinguanchedarias.owgejava.pojo.UnitWithRequirementInformation;
 import com.kevinguanchedarias.owgejava.responses.CriticalAttackInformationResponse;
 import lombok.AllArgsConstructor;
-import org.hibernate.Hibernate;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -56,8 +52,6 @@ public class UnitRestService implements SyncSource {
     private final FactionBo factionBo;
     private final UnitBo unitBo;
     private final CriticalAttackBo criticalAttackBo;
-    private final UnitTypeInheritanceFinderService unitTypeInheritanceFinderService;
-    private final SpeedImpactGroupBo speedImpactGroupBo;
 
     /**
      * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
@@ -118,36 +112,14 @@ public class UnitRestService implements SyncSource {
     private List<UnitDto> findUnlocked(UserStorage user) {
         List<Unit> units = unlockedRelationBo.unboxToTargetEntity(
                 unlockedRelationBo.findByUserIdAndObjectType(user.getId(), ObjectEnum.UNIT));
-
-        units.forEach(current -> Hibernate.initialize(current.getInterceptableSpeedGroups()));
-        var convert = new UnitDto();
-        return convert.dtoFromEntity(UnitDto.class, units);
+        var ous = units.stream()
+                .map(unit -> ObtainedUnit.builder().unit(unit).user(user).build())
+                .toList();
+        return obtainedUnitBo.findCompletedAsDto(user, ous).stream().map(ObtainedUnitDto::getUnit).toList();
     }
 
     private List<ObtainedUnitDto> findInMyPlanets(UserStorage user) {
-        List<ObtainedUnit> entities = obtainedUnitBo.findDeployedInUserOwnedPlanets(user.getId());
-
-        entities.stream()
-                .filter(ou -> ou.getUnit().getSpeedImpactGroup() != null)
-                .map(ObtainedUnit::getUnit)
-                .forEach(unit -> {
-                    Hibernate.initialize(unit.getSpeedImpactGroup());
-                    unit.getSpeedImpactGroup().setRequirementGroups(null);
-                });
-        entities.forEach(current -> Hibernate.initialize(current.getUnit().getInterceptableSpeedGroups()));
-        var result = obtainedUnitBo.toDto(entities);
-        result.stream()
-                .map(ObtainedUnitDto::getUnit)
-                .filter(unit -> unit.getSpeedImpactGroup() == null)
-                .forEach(unitWithNullSpeedImpact -> unitWithNullSpeedImpact.setSpeedImpactGroup(
-                        speedImpactGroupBo.toDto(
-                                unitTypeInheritanceFinderService.findUnitTypeMatchingCondition(
-                                        unitWithNullSpeedImpact.getTypeId(),
-                                        unitType -> unitType.getSpeedImpactGroup() != null
-                                ).map(UnitType::getSpeedImpactGroup).orElse(null)
-                        )
-                ));
-        return result;
+        return obtainedUnitBo.findCompletedAsDto(user);
     }
 
     private UserStorage findLoggedInUser() {